+++ src/BitwardenCli.Core/Serialization/BitwardenJsonContext.cs
@@ -4,5 +4,9 @@ using BitwardenCli.Core.Models;
 namespace BitwardenCli.Core.Serialization;
 
 [JsonSerializable(typeof(BitwardenStatus))]
+[JsonSerializable(typeof(VaultItem))]
+[JsonSerializable(typeof(VaultItem[]))]
+[JsonSerializable(typeof(VaultFolder))]
+[JsonSerializable(typeof(VaultFolder[]))]
 [JsonSourceGenerationOptions(PropertyNameCaseInsensitive = true)]
 internal sealed partial class BitwardenJsonContext : JsonSerializerContext;